+++ example/lib/main.dart
@@ -30,6 +30,7 @@ import 'package:flutter_widgetz_example/pages/search_bar_page.dart';
 import 'package:flutter_widgetz_example/pages/slider_page.dart';
 import 'package:flutter_widgetz_example/pages/sliver_grid_page.dart';
 import 'package:flutter_widgetz_example/pages/sliver_list_page.dart';
+import 'package:flutter_widgetz_example/pages/text_field_page.dart';
 import 'package:flutter_widgetz_example/pages/time_field_page.dart';
 
 /// To run the example app, clone/fork the repo!
@@ -194,5 +195,6 @@ const List<PageInfo> _pages = <PageInfo>[
   PageInfo(Icons.tune, SliderPage(), 'Sliders'),
   PageInfo(Icons.grid_4x4, SliverGridPage(), 'Sliver Grid'),
   PageInfo(Icons.list, SliverListPage(), 'Sliver List'),
+  PageInfo(Icons.text_fields, TextFieldPage(), 'Text Field'),
   PageInfo(Icons.timelapse, TimeFieldPage(), 'Time Field'),
 ];